Extract a class from Conman to add a contact.
