+++ bin/app
@@ -5,8 +5,10 @@ $LOAD_PATH.unshift File.join(File.dirname(__FILE__), '..', 'lib')
 require 'console'
 require 'ui'
 require 'db'
+require 'creator'
 require 'conman'
 
-console = Console.new($stdin, $stdout)
-conman  = Conman.new(UI.new(console), DB.new)
+ui      = UI.new(Console.new($stdin, $stdout))
+db      = DB.new
+conman  = Conman.new(ui, db, Creator.new(ui, db))
 conman.run
